separate class for spindle config and test for sql connection for ui
